Rename `docker rebuild` to `docker rebase`
